Add main.ejs and forum-teacher.ejs
Main.ejs is the first page that the user will see after he/she connects to the website and it has links to the forum and professor pages. The forum-teacher.ejs page is the forum page for someone who has administrative access so they can make a new post that will be sent to the database.